+++ xcore/kernel/runtime/middlewares/retry.py
@@ -8,7 +8,7 @@ import asyncio
 import logging
 from typing import Callable
 
-from ..middleware import Middleware
+from .middleware import Middleware
 
 logger = logging.getLogger("xcore.runtime.middleware.retry")
